+++ python/paddle/fluid/tests/unittests/test_sgd_op.py
@@ -214,7 +214,7 @@ class TestSGDV2(unittest.TestCase):
         paddle.disable_static()
         value = np.arange(26).reshape(2, 13).astype("float32")
         a = paddle.to_tensor(value)
-        linear = paddle.nn.Linear(13, 5, dtype="float32")
+        linear = paddle.nn.Linear(13, 5)
         # This can be any optimizer supported by dygraph.
         adam = paddle.optimizer.SGD(learning_rate=0.01,
                                     parameters=linear.parameters(),
